+++ apps/server/CHANGELOG.md
@@ -1,5 +1,16 @@
 # @ying-tunnel/server
 
+## 1.1.0
+
+### Minor Changes
+
+- custom implementation of random key function
+
+### Patch Changes
+
+- Updated dependencies
+  - @ying-tunnel/lib@1.1.0
+
 ## 1.0.2
 
 ### Patch Changes