+++ packages/lib/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @ying-tunnel/lib
 
+## 1.1.0
+
+### Minor Changes
+
+- custom implementation of random key function
+
 ## 1.0.1
 
 ### Patch Changes